Create the struct of the HomeViewWithPets
